% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_sex_table}
\alias{age_sex_table}
\title{Age-sex indicator table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 14 rows and 4 columns.
}
\usage{
age_sex_table
}
\description{
Age-sex indicator table
}
\keyword{datasets}
