% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_utils.R
\name{bq_append_raw_data}
\alias{bq_append_raw_data}
\title{Append raw data to BigQuery}
\usage{
bq_append_raw_data(dt, con = NULL, frequency = "monthly", chunk_size = 500000L)
}
\arguments{
\item{dt}{data.table to append (must have: dataElement, period, orgUnit,
value, timestamp, version)}

\item{con}{A DBI BigQuery connection (or NULL to create one)}

\item{frequency}{Either \code{"weekly"} or \code{"monthly"}}

\item{chunk_size}{Upper limit on download chunk}
}
\description{
Append raw data to BigQuery
}
