% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_utils.R
\name{bq_append_version_metadata}
\alias{bq_append_version_metadata}
\title{Write version metadata to BigQuery}
\usage{
bq_append_version_metadata(version_df, con = NULL, frequency = "monthly")
}
\arguments{
\item{version_df}{data.frame with version metadata}

\item{con}{A DBI BigQuery connection (or NULL to create one)}

\item{frequency}{Either \code{"weekly"} or \code{"monthly"}}
}
\description{
Write version metadata to BigQuery
}
