% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_utils.R
\name{bq_get_latest_version}
\alias{bq_get_latest_version}
\title{Get the latest version number from BigQuery}
\usage{
bq_get_latest_version(con = NULL, frequency = "monthly")
}
\arguments{
\item{con}{A DBI BigQuery connection (or NULL to create one)}

\item{frequency}{Either \code{"weekly"} or \code{"monthly"}}
}
\value{
Integer value of the latest version
}
\description{
Get the latest version number from BigQuery
}
