% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_utils.R
\name{bq_get_outbreak_data}
\alias{bq_get_outbreak_data}
\title{Read pre-computed outbreak index data from BigQuery}
\usage{
bq_get_outbreak_data(con = NULL)
}
\arguments{
\item{con}{A DBI BigQuery connection (or NULL to create one)}
}
\value{
data.table with outbreak index data
}
\description{
Read pre-computed outbreak index data from BigQuery
}
