% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_utils.R
\name{bq_get_version_metadata}
\alias{bq_get_version_metadata}
\title{Get version metadata from BigQuery}
\usage{
bq_get_version_metadata(con = NULL, frequency = "monthly", version_id = NULL)
}
\arguments{
\item{con}{A DBI BigQuery connection (or NULL to create one)}

\item{frequency}{Either \code{"weekly"} or \code{"monthly"}}

\item{version_id}{Version to retrieve (NULL = latest)}
}
\value{
data.table with version metadata
}
\description{
Get version metadata from BigQuery
}
