% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhis_indicator_table}
\alias{dhis_indicator_table}
\title{DHIS indicator table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 71 rows and 7 columns.
}
\usage{
dhis_indicator_table
}
\description{
DHIS indicator table
}
\keyword{datasets}
