% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_pop}
\alias{district_pop}
\title{District population data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 146 rows and 2 columns.
}
\usage{
district_pop
}
\description{
District population data
}
\keyword{datasets}
