% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_data}
\alias{get_data}
\title{Get the data from a database with the option to subset to a specified version and id values}
\usage{
get_data(db_path, id_list = NULL, version_id = NULL)
}
\arguments{
\item{db_path}{Location of the database}

\item{id_list}{A list of id names and values, if NULL, pulls full table}

\item{version_id}{Integer version number, if NULL, pulls latest version}
}
\description{
Get the data from a database with the option to subset to a specified version and id values
}
