% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_db_diff}
\alias{get_db_diff}
\title{Get the data from new data that differs from what is present in the database}
\usage{
get_db_diff(new_data, db_path)
}
\arguments{
\item{new_data}{data.frame with new data for comparison to the data in the database}

\item{db_path}{Location of the database}
}
\value{
data.frame with the data from new data that differs from what is present in the database
}
\description{
Get the data from new data that differs from what is present in the database
}
