% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_id_vars}
\alias{get_id_vars}
\title{Get the names of the variables that uniquely identify an observation}
\usage{
get_id_vars(db_path)
}
\arguments{
\item{db_path}{Location of the database}
}
\value{
Vector of column names that uniquely identify an observation
}
\description{
Get the names of the variables that uniquely identify an observation
}
