% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_latest_version}
\alias{get_latest_version}
\title{Get the latest version number from the database}
\usage{
get_latest_version(db_path)
}
\arguments{
\item{db_path}{Location of the database}
}
\value{
Integer value of latest version
}
\description{
Get the latest version number from the database
}
