% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{get_period_range}
\alias{get_period_range}
\title{Get the a vector of periods follow DHIS formatting that spans that start and end period provided}
\usage{
get_period_range(frequency, year_start, sub_year_start, year_end, sub_year_end)
}
\arguments{
\item{frequency}{Text of either weekly or monthly}

\item{year_start}{Integer for start year}

\item{sub_year_start}{Integer for either week or month of the year to start}

\item{year_end}{Integer for end year}

\item{sub_year_end}{Integer for either week or month of the year to end}
}
\value{
A vector of periods following the DHIS formatting standard of either YYYYW{week number with no leading zero} or YYYY{month number with a leading zero for single digit integers}
}
\description{
Get the a vector of periods follow DHIS formatting that spans that start and end period provided
}
