% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_value_var}
\alias{get_value_var}
\title{Get the name of the value column}
\usage{
get_value_var(db_path)
}
\arguments{
\item{db_path}{Location of the database}
}
\value{
Column name for the value
}
\description{
Get the name of the value column
}
