% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_version_metadata}
\alias{get_version_metadata}
\title{Get metadata for a specific version}
\usage{
get_version_metadata(db_path, version_id = NULL)
}
\arguments{
\item{db_path}{Location of the database}

\item{version_id}{Integer for version number to pull metadata for}
}
\value{
Integer value of latest version
}
\description{
Get metadata for a specific version
}
