% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_facility_table}
\alias{health_facility_table}
\title{Health facility table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8676 rows and 22 columns.
}
\usage{
health_facility_table
}
\description{
Health facility table
}
\keyword{datasets}
