% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_table}
\alias{indicator_table}
\title{DHIS indicator table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 87 rows and 6 columns.
}
\usage{
indicator_table
}
\description{
DHIS indicator table
}
\keyword{datasets}
