% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{make_human_readable}
\alias{make_human_readable}
\title{Merge on human readable columns to raw data}
\usage{
make_human_readable(dt)
}
\arguments{
\item{dt}{data.table with raw data in it}
}
\value{
data.table with human readable columns merged on
}
\description{
Merge on human readable columns to raw data
}
