% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{make_month_map}
\alias{make_month_map}
\title{Make a table that provides the date information associated with a DHIS period}
\usage{
make_month_map(min_date = "2013-01-01")
}
\arguments{
\item{min_date}{Date string in YYYY-MM-DD format for start date of the map}
}
\value{
A table with the period in DHIS format, year, month, date_start, date_mid, and date_end
}
\description{
Make a table that provides the date information associated with a DHIS period
}
