% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{make_week_map}
\alias{make_week_map}
\title{Make a table that provides the date information associated with a DHIS period}
\usage{
make_week_map(min_date = "2012-12-31")
}
\arguments{
\item{min_date}{Date string in YYYY-MM-DD format for start date of the map}
}
\value{
A table with the period in DHIS format, year, week, date_start, date_mid, and date_end
}
\description{
Make a table that provides the date information associated with a DHIS period
}
