% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{old_dhis_indicator_table}
\alias{old_dhis_indicator_table}
\title{Old DHIS indicator table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 5 columns.
}
\usage{
old_dhis_indicator_table
}
\description{
Old DHIS indicator table
}
\keyword{datasets}
