% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_district_shp}
\alias{uga_district_shp}
\title{District shapefile}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 146 rows and 3 columns.
}
\usage{
uga_district_shp
}
\description{
District shapefile
}
\keyword{datasets}
