% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_region_shp}
\alias{uga_region_shp}
\title{Region shapefile version1.0}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
uga_region_shp
}
\description{
Region shapefile version1.0
}
\keyword{datasets}
