% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_facility_table}
\alias{health_facility_table}
\title{Health Facility and Administrative Units Table}
\format{
A data frame with 8676 rows and 22 columns:
\describe{
  \item{location_name}{Name of Health facility}
  \item{location_id}{DHIS2 assigned ID for health facility}
  \item{parent_id}{DHIS2 assigned ID of parent health facility (subcounty)}
  \item{level}{DHIS2 assigned hierarchical level}
  \item{authority}{The governing body or authority responsible for the facility}
  \item{facility_level}{The original classification level of the health facility}
  \item{medical_bureaus}{The medical bureau affiliation (e.g., UPMB, UCMB)}
  \item{operational_status}{Current status of the facility (e.g., Functional, Non-Functional)}
  \item{ownership}{Type of ownership (e.g., Public, Private, PNFP)}
  \item{private_facilities}{Indicator for private sector classification (PFP, PNFP)}
  \item{public_facilities}{Indicator for public sector classification (e.g. MOH,BOU, Local Government)}
  \item{reporting_status}{Current status of data reporting compliance}
  \item{path_to_top_parent}{String of DHIS2 IDs from current health facility to the top-level parent}
  \item{region_name}{Name of region associated with health facility}
  \item{district_name}{Name of district associated with health facility}
  \item{DLG_name}{Name of District Local Government (DLG) associated with health facility}
  \item{subcounty_name}{Name of subcounty associated with health facility}
  \item{latitude}{Geographic coordinate: North-South position}
  \item{longitude}{Geographic coordinate: East-West position}
  \item{geom_enhanced}{Indicator if geometry was added through external web-scraping and not present in DHIS2}
  \item{facility_level_2}{Secondary facility level classification (clinic, hospital, drugshop, other)}
  \item{facility_type}{Secondary facility label created to match other surveys on health seeking (private, gov_HC, private_hosp, drugshop, gov_hosp, other)}
}
}
\usage{
health_facility_table
}
\description{
A comprehensive list of health facilities and administrative hierarchies,
including ownership, operational status, and geographic coordinates.
}
\keyword{datasets}
