% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_table}
\alias{loc_table}
\title{Location hierarchy table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10476 rows and 16 columns.
}
\usage{
loc_table
}
\description{
Location hierarchy table
}
\keyword{datasets}
