% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_subcounty_shp}
\alias{uga_subcounty_shp}
\title{Subcounty shapefile version1.0}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2206 rows and 4 columns.
}
\usage{
uga_subcounty_shp
}
\description{
Subcounty shapefile version1.0
}
\keyword{datasets}
